fix deepseek image test env import
